trace: fix unused warnings with NDEBUG
When pppoat is built with NDEBUG configuration, asserts and debug
logs must be built into the binary. It was achieved by replacing
the macros with "do {} while (0)". But this construction ignores
arguments and if the arguments are not used anywhere else then
compiler generates a warning.

Fix the above issue by using the following pattern:
	if (!NDEBUG)
		func_implementation();

Compiler with enabled optimizations won't built such an unreachable
code, but from other side, all arguments will be used.

Also, move log wrappers such as pppoat_debug() to trace.h.